rename example file
