+++ examples/component_svg_viewer.rs
@@ -1,0 +1,23 @@
+#![cfg_attr(
+    all(not(debug_assertions), target_os = "windows"),
+    windows_subsystem = "windows"
+)]
+
+use freya::prelude::*;
+
+fn main() {
+    launch(LaunchConfig::new().with_window(WindowConfig::new(app)))
+}
+
+fn app() -> impl IntoElement {
+    rect().expanded().center().child(
+        SvgViewer::new(r#"<!-- https://github.com/bbfh-dev/eyecons/blob/main/.docs/pin.svg?short_path=6659513 -->
+        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16" fill="currentColor">
+            <path
+                d="m8 1.9365e-7a1.7921 1.7921 0 0 0-0.89522 0.24048l-5.3774 3.104a1.7921 1.7921 0 0 0-0.89522 1.552l2e-8 6.208a1.7921 1.7921 0 0 0 0.89522 1.552l5.3774 3.104a1.7921 1.7921 0 0 0 1.7904 0l5.3774-3.104a1.7921 1.7921 0 0 0 0.89522-1.552v-6.208a1.7921 1.7921 0 0 0-0.89522-1.552l-5.3774-3.104a1.7921 1.7921 0 0 0-0.89522-0.24048zm0 1.7925 4.3407 2.5058-4.3407 2.5058-4.3407-2.5058zm-5.3754 4.3003 4.3407 2.5058v5.0116l-4.3407-2.5058zm10.751 0v5.0116l-4.3407 2.5058 1e-7 -5.0116z" />
+                </svg>"#
+                .as_bytes())
+            .width(Size::px(300.))
+            .height(Size::px(300.)).color(Color::BLACK),
+    )
+}
